+++ secrets/nomad/cmd/main.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/JanMa/openbao-plugin-secrets-nomad/nomad"
 	hclog "github.com/hashicorp/go-hclog"
-	"github.com/hashicorp/vault/api"
-	"github.com/hashicorp/vault/sdk/plugin"
+	"github.com/openbao/openbao/api"
+	"github.com/openbao/openbao/sdk/plugin"
 )
 
 func main() {