+++ secrets/nomad/path_config_access.go
@@ -1,4 +1,5 @@
 // Copyright (c) HashiCorp, Inc.
+// Copyright (c) Jan Martens <jan@martens.eu.org>
 // SPDX-License-Identifier: MPL-2.0
 
 package nomad
@@ -7,8 +8,8 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/hashicorp/vault/sdk/framework"
-	"github.com/hashicorp/vault/sdk/logical"
+	"github.com/openbao/openbao/sdk/framework"
+	"github.com/openbao/openbao/sdk/logical"
 )
 
 const configAccessKey = "config/access"